+++ kameleoon/client/src/Data/Personalization.php
@@ -2,16 +2,20 @@
 
 namespace Kameleoon\Data;
 
+use Kameleoon\Helpers\TimeHelper;
+
 /** @internal */
 class Personalization implements BaseData
 {
     private int $id;
     private int $variationId;
+    private int $assignmentDateMillis;
 
-    public function __construct(int $id, int $variationId)
+    public function __construct(int $id, int $variationId, ?int $assignmentDateMillis = null)
     {
         $this->id = $id;
         $this->variationId = $variationId;
+        $this->assignmentDateMillis = $assignmentDateMillis ?? TimeHelper::nowInMilliseconds();
     }
 
     public function getId(): int
@@ -24,8 +28,16 @@ class Personalization implements BaseData
         return $this->variationId;
     }
 
+    public function getAssignmentDateMillis(): int
+    {
+        return $this->assignmentDateMillis;
+    }
+
     public function __toString(): string
     {
-        return "Personalization{id:" . $this->id . ",variationId:" . $this->variationId . "}";
+        return "Personalization{id:" . $this->id .
+            ",variationId:" . $this->variationId .
+            ",assignmentDateMillis:" . $this->assignmentDateMillis .
+            "}";
     }
 }