+++ kameleoon/client/src/Targeting/Condition/TargetExperimentCondition.php
@@ -6,7 +6,7 @@ namespace Kameleoon\Targeting\Condition;
 
 use Kameleoon\Logging\KameleoonLogger;
 
-class TargetExperimentCondition extends TargetingCondition
+class TargetExperimentCondition extends VisitorScopeCondition
 {
     const TYPE = "TARGET_EXPERIMENT";
 
@@ -16,7 +16,7 @@ class TargetExperimentCondition extends TargetingCondition
 
     public function __construct($conditionData)
     {
-        parent::__construct($conditionData);
+        parent::__construct($conditionData, VisitorScopeCondition::VISIT_SCOPE_CURRENT_VISIT);
         $this->variationId = $conditionData->variationId ?? -1;
         $this->experimentId = $conditionData->experimentId ?? -1;
         $this->variationMatchType = $conditionData->variationMatchType ?? TargetingOperator::UNKNOWN;
@@ -27,16 +27,22 @@ class TargetExperimentCondition extends TargetingCondition
         if (!is_array($data)) {
             return false;
         }
-        $variations = $data;
+        $visitorVisits = $data["visitorVisits"] ?? null;
+        $variations = $data["variations"] ?? [];
         $variation = $variations[$this->experimentId] ?? null;
+        $variationExists = $variation !== null
+            && $variation->getAssignmentDateMillis() >= $this->getAssignmentThresholdMillis($visitorVisits);
         switch ($this->variationMatchType) {
             case TargetingOperator::ANY:
-                return $variation !== null;
+                return $variationExists;
             case TargetingOperator::EXACT:
-                return ($variation !== null) && ($variation->getVariationId() == $this->variationId);
+                return $variationExists && $variation->getVariationId() == $this->variationId;
+            default:
+                break;
         }
         KameleoonLogger::error(
-            "Unexpected variation match type for 'TargetExperimentCondition' condition: '%s'", $this->variationMatchType
+            "Unexpected variation match type for 'TargetExperimentCondition' condition: '%s'",
+            $this->variationMatchType
         );
         return false;
     }